% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDiscrPath.R
\name{allDiscrPath}
\alias{allDiscrPath}
\title{Finds all discriminating paths for R4}
\usage{
allDiscrPath(pag, start_inter, a, b, c, count, start_stat_edges)
}
\description{
Finds all discriminating paths for R4
}

