% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allUncovPath.R
\name{allUncovPath}
\alias{allUncovPath}
\title{Find all uncovered paths}
\usage{
allUncovPath(p, pag, a, b, c, suffStat, sepset, mix_modeling, mix_mat)
}
\arguments{
\item{p}{Number of variables}

\item{pag}{The graph in matrix form}

\item{a}{A vertex}

\item{b}{A vertex}

\item{c}{A vertex}
}
\value{
List containing all paths and associated mix_mat 3D arrays
}
\description{
Find all uncovered paths
}

