% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEdges1.R
\name{checkEdges1}
\alias{checkEdges1}
\title{Checks conditional dependence and mixtures for R4}
\usage{
checkEdges1(suffStat, indepTest, alpha, apag, sepset, path, mix_modeling,
  mix_mat)
}
\description{
Checks conditional dependence and mixtures for R4
}

