% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_triple.R
\name{dep.triple}
\alias{dep.triple}
\title{Find triples with dependent edges for RFCI}
\usage{
dep.triple(suffStat, indepTest, alpha, sepset, apag, unshTripl, unshVect,
  trueVstruct, verbose = FALSE)
}
\description{
Find triples with dependent edges for RFCI
}

