% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_mix_modeling.R
\name{gauss_mix_modeling}
\alias{gauss_mix_modeling}
\title{Gaussian conditional mixture modeling wrapper}
\usage{
gauss_mix_modeling(a, b, cond_set, suffStat, joint)
}
\arguments{
\item{a}{First non-conditioning variable number}

\item{b}{Second non-conditioning variable number}

\item{cond_set}{Conditioning set variable numbers}

\item{joint}{Set to TRUE or FALSE for bivariate or univariate conditional mixture modeling, respectively}
}
\value{
mix_res The mixture results; a positive integer in the bivariate case, or a list with two elements in the univariate case
}
\description{
Gaussian conditional mixture modeling wrapper
}

