% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_mix_modeling_joint.R
\name{gauss_mix_modeling_joint}
\alias{gauss_mix_modeling_joint}
\title{Joint (bivariate) Gaussian conditional mixture modeling}
\usage{
gauss_mix_modeling_joint(a, b, cond_set, suffStat)
}
\arguments{
\item{a}{First non-conditioning variable number}

\item{b}{Second non-conditioning variable number}

\item{cond_set}{Conditioning set variable numbers}
}
\value{
mix_res Single positive integer specifying number of mixtures for p(a,b|cond_set)
}
\description{
Joint (bivariate) Gaussian conditional mixture modeling
}

