% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_mix_modeling_single.R
\name{gauss_mix_modeling_single}
\alias{gauss_mix_modeling_single}
\title{Univariate Gaussian conditional mixture modeling}
\usage{
gauss_mix_modeling_single(a, b, cond_set, suffStat)
}
\arguments{
\item{a}{First non-conditioning variable number}

\item{b}{Second non-conditioning variable number}

\item{cond_set}{Conditioning set variable numbers}
}
\value{
mix_res A list containing two elements. The first element is a single positive integer for the number of mixtures for p(a|cond_set); likewise for the second element for p(b|cond_set).
}
\description{
Univariate Gaussian conditional mixture modeling
}

