% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_discovery.R
\name{IP_discovery}
\alias{IP_discovery}
\title{RFCI's v-structure discovery procedure adapted for F2CI}
\usage{
IP_discovery(suffStat, indepTest, alpha, max.cs)
}
\arguments{
\item{suffStat}{List containing suffStat$data}

\item{indepTest}{The independence test}

\item{alpha}{Type I error rate for indepTest}
}
\value{
List containing
}
\description{
RFCI's v-structure discovery procedure adapted for F2CI
}

