% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_discovery2.R
\name{OR_discovery2}
\alias{OR_discovery2}
\title{Applies F2CI's orientation rules}
\usage{
OR_discovery2(apag, suffStat, indepTest, alpha, mix_modeling, sepset, count,
  rules = rep(TRUE, 10), verbose = FALSE)
}
\description{
Applies F2CI's orientation rules
}

