% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Repeat rows or columns of a vector or matrix}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{Vector or matrix}

\item{m}{number of row copies}

\item{n}{number of column copies}
}
\value{
Matrix.
}
\description{
Repeat rows or columns of a vector or matrix
}
\examples{
repmat(matrix(rnorm(10*2),ncol=2),2,2)
}

