% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCI_vstruc.R
\name{RFCI_vstruc}
\alias{RFCI_vstruc}
\title{RFCI's v-structure discovery procedure adapted for F2CI}
\usage{
RFCI_vstruc(suffStat, indepTest, alpha, sepset, g.amat, unshTripl, unshVect,
  mix_modeling, verbose = FALSE)
}
\arguments{
\item{suffStat}{List containing suffStat$data}

\item{indepTest}{The independence test}

\item{alpha}{Type I error rate for indepTest}

\item{sepset}{Matrix of separating sets}

\item{g.amat}{Graph returned from skeleton discovery procedure}
}
\value{
List containing separating sets (sepset), n
}
\description{
RFCI's v-structure discovery procedure adapted for F2CI
}

