% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateList_m.R
\name{updateList_m}
\alias{updateList_m}
\title{Update list function used in allUncovPath function}
\usage{
updateList_m(path, set, old.list)
}
\arguments{
\item{path}{The path to update}

\item{set}{Set of vertices to update path with}

\item{path}{old.list The old list of paths}
}
\value{
List containing all updated paths
}
\description{
Update list function used in allUncovPath function
}

